### 2016-09-05 #73 @cvitolo
hddtools package
This R package is an open source project designed to facilitate non-programmatic access to a variety of online open data sources: the Global Runoff Data Center (GRDC), the Scottish Environment Protection Agency (SEPA), the Top-Down modelling Working Group (Data60UK and MOPEX), Met Office Hadley Centre Observation Data (HadUKP Data) and NASA's Tropical Rainfall Measuring Mission (TRMM).

```
Package: hddtools
Title: Hydrological Data Discovery Tools
Version: 0.3.0
Authors@R: c(person("Claudia", "Vitolo", email = "cvitolodev@gmail.com", role = c("aut", "cre")))
Maintainer: Claudia Vitolo <cvitolodev@gmail.com>
Description: Facilitates discovery and handling of hydrological data, non-programmatic access to catalogues and databases.
Imports: zoo, sp, RCurl, XML, rnrfa, Hmisc, raster, stringr
Suggests: testthat
Repository: CRAN
```

https://github.com/cvitolo/hddtools/

Hydrologists, environmental scientists and practitioners.

    http://dx.doi.org/10.5281/zenodo.61570

@masalmon
