### 2016-09-05 a594fd7
Added: vignette + JOSS paper + document to review the package under the ropensci project
