+++ inst/paper/paper.bib
@@ -1,0 +1,40 @@
+@misc{hddtools-archive,
+  author       = {Claudia Vitolo},
+  title        = {hddtools: Hydrological Data Discovery Tools},
+  month        = Sep,
+  year         = 2016,
+  doi          = {10.5281/zenodo.61570},
+  url          = {http://dx.doi.org/10.5281/zenodo.61570}
+}
+
+@Manual{R-base,
+    title = {R: A Language and Environment for Statistical Computing},
+    author = {{R Core Team}},
+    organization = {R Foundation for Statistical Computing},
+    address = {Vienna, Austria},
+    year = {2016},
+    url = {https://www.R-project.org/},
+}
+
+@article{rdefra,
+  doi = {10.21105/joss.00051},
+  url = {http://dx.doi.org/10.21105/joss.00051},
+  year  = {2016},
+  month = {aug},
+  publisher = {The Open Journal},
+  volume = {1},
+  number = {4},
+  author = {Claudia Vitolo and Andrew Russell and Allan Tucker},
+  title = {rdefra: Interact with the {UK} {AIR} Pollution Database from {DEFRA}},
+  journal = {{JOSS}}
+}
+
+@Manual{rnrfa,
+  author = {Claudia Vitolo and
+            Matthew Fry and
+            Wouter Buytaert},
+  title = {rnrfa: UK National River Flow Archive Data from R},
+  year = 2015,
+  note = {R package version 0.5.3},
+  url = {https://CRAN.R-project.org/package=rnrfa}
+}
