### 2016-09-06 #73 @noamross
### Editor checks:

- [x] **Fit**: The package meets or criteria [fit](policies.md#fit) and [overlap](policies.md#fit)

---

#### Editor comments

Thank you for the submission, @cvitolo!  Before we progress to finding reviewers, we would like the package to have a test suite with a set of tests that cover a good portion of the code. I see unit tests mentioned in your NEWS file - perhaps they're just not committed for some reason?

---
In addition, I ran `goodpractice::gp()`, which we are starting to do to flag some things.  Here are the results, which  it may be helpful to get started with before review:

```
── GP hddtools ──────────────

It is good practice to

  ✖ write unit tests for all functions, and all package code in general. At least some lines are not covered in this package.

    R/bboxSpatialPolygon.R:23:NA
    R/bboxSpatialPolygon.R:24:NA
    R/bboxSpatialPolygon.R:25:NA
    R/bboxSpatialPolygon.R:26:NA
    R/bboxSpatialPolygon.R:27:NA
    ... and 768 more lines

  ✖ omit "Date" in DESCRIPTION. It is not required and it gets invalid quite often. A build date will be added to the package
    when you perform `R CMD build` on it.
  ✖ use '<-' for assignment instead of '='. '<-' is the standard, and R users and developers are used it and it is easier to
    read your code for them if you use '<-'.

    R/GRDC.R:188:11

  ✖ avoid long code lines, it is bad for readability. Also, many people prefer editor windows that are about 80 characters

    R/Data60UK.R:71:1
    R/Data60UK.R:78:1
    R/GRDC.R:194:1
    R/GRDC.R:196:1
    R/GRDC.R:213:1
    ... and 25 more lines

  ✖ avoid calling setwd(), it changes the global environment. If you need it, consider using on.exit() to restore the working
    directory.

    R/TRMM.R:55:3

  ✖ avoid sapply(), it is not type safe. It might return a vector, or a list, depending on the input data. Consider using
    vapply() instead.

    R/getContent.R:23:9

  ✖ fix this R CMD check error/warning/note: Found the following hidden files and directories: .travis.yml These were most
    likely included in error. See section ‘Package structure’ in the ‘Writing R Extensions’ manual.  [Editor's note: this file should be in .Rbuildignore]
  ✖ avoid 'T' and 'F', as they are just variables which are set to the logicals 'TRUE' and 'FALSE' by default, but are not
    reserved words and hence can be overwritten by the user.  Hence, one should always use 'TRUE' and 'FALSE' for the
    logicals.

    R/GRDC.R:NA:NA
    R/HadDAILY.R:NA:NA
```
