### 2016-09-06 444682a
Added unit tests
