### 2016-09-06 acd115f
Fixed problems highlighted by goodpractice::gp()
