+++ R/GRDC.R
@@ -20,7 +20,8 @@
 #'   catalogueGRDC()
 #'
 #'   # Define a bounding box
-#'   bbox <- list(lonMin=-3.82,latMin=52.41,lonMax=-3.63,latMax=52.52)
+#'   bbox <- list(lonMin = -3.82, latMin = 52.41,
+#'                lonMax = -3.63, latMax = 52.52)
 #'
 #'   # Filter the catalogue
 #'   catalogueGRDC(bbox)
@@ -28,16 +29,16 @@
 #'
 
 catalogueGRDC <- function(bbox = NULL, stationID = NULL,
-                           metadataColumn=NULL, entryValue=NULL,
-                           mdDescription=FALSE){
+                          metadataColumn = NULL, entryValue = NULL,
+                          mdDescription = FALSE){
 
   # Retrieve the catalogue
   temp <- system.file("extdata/GRDC/GRDC_Stations_20140320.csv",
                       package = 'hddtools')
-  grdcAll <- read.csv(temp,sep=",")
+  grdcAll <- read.csv(temp,sep = ",")
 
   if (!is.null(stationID)) {
-    grdcSelected <- subset(grdcAll, (grdcAll$grdc_no==stationID) )
+    grdcSelected <- subset(grdcAll, (grdcAll$grdc_no == stationID) )
   }else{
     grdcSelected <- grdcAll
   }
@@ -88,10 +89,10 @@ catalogueGRDC <- function(bbox = NULL, stationID = NULL,
   names(grdcTable)[9]  <- "Latitude"
   names(grdcTable)[10] <- "Longitude"
 
-  if (mdDescription==TRUE){
+  if (mdDescription == TRUE){
 
     temp <- system.file("extdata/GRDC/GRDC_legend.csv", package = 'hddtools')
-    grdcLegend <- read.csv(temp, header=F)
+    grdcLegend <- read.csv(temp, header = FALSE)
 
     grdcTable <- cbind(grdcLegend,t(grdcTable))
     row.names(grdcTable) <- NULL
@@ -146,46 +147,46 @@ catalogueGRDC <- function(bbox = NULL, stationID = NULL,
 #'
 #' @examples
 #' \dontrun{
-#'   x <- tsGRDC(stationID=1107700)
+#'   x <- tsGRDC(stationID = 1107700)
 #' }
 #'
 
-tsGRDC <- function(stationID, plotOption=FALSE){
+tsGRDC <- function(stationID, plotOption = FALSE){
 
-  options(warn=-1)
+  options(warn = -1)
 
   temp <- catalogueGRDC()
 
-  if ( temp[which(temp$grdc_no==stationID),"statistics"] == 1 ){
+  if ( temp[which(temp$grdc_no == stationID),"statistics"] == 1 ){
 
     # Retrieve look-up table
     temp <- system.file("extdata/GRDC/GRDC_LTMMD.csv", package = 'hddtools')
-    grdcLTMMD <- read.csv(temp,sep="\t")
+    grdcLTMMD <- read.csv(temp, sep = "\t")
 
     # Retrieve WMO region from catalogue
     wmoRegion <- catalogueGRDC(stationID = stationID)$wmo_reg
 
     # Retrieve ftp server location
-    zipFile <- as.character(grdcLTMMD[which(grdcLTMMD$WMO.Region==wmoRegion),
+    zipFile <- as.character(grdcLTMMD[which(grdcLTMMD$WMO.Region == wmoRegion),
                                       "Archive"])
 
     # create a temporary directory
     td <- tempdir()
 
     # create the placeholder file
-    tf <- tempfile(tmpdir=td, fileext=".zip")
+    tf <- tempfile(tmpdir = td, fileext = ".zip")
 
[…]
