+++ tests/testthat/test-KGClimateClass.R
@@ -9,7 +9,7 @@ test_that("Test KGClimateClass function", {
   # Get climate classes
   x <- KGClimateClass(bbox)
 
-  expect_that("data.frame" %in% class(x), equals(TRUE))
+  expect_that(class(x) == "data.frame", equals(TRUE))
   expect_that(all(dim(x) == c(1, 3)), equals(TRUE))
   expect_that(all(x$ID == 15, x$Class == "Cfb", x$Frequency == 2), equals(TRUE))
 
