+++ man/catalogueGRDC.Rd
@@ -30,7 +30,8 @@ This function interfaces the Global Runoff Data Centre database which provides r
   catalogueGRDC()
 
   # Define a bounding box
-  bbox <- list(lonMin=-3.82,latMin=52.41,lonMax=-3.63,latMax=52.52)
+  bbox <- list(lonMin = -3.82, latMin = 52.41,
+               lonMax = -3.63, latMax = 52.52)
 
   # Filter the catalogue
   catalogueGRDC(bbox)
