+++ DESCRIPTION
@@ -2,7 +2,6 @@ Package: hddtools
 Type: Package
 Title: Hydrological Data Discovery Tools
 Version: 0.3.0
-Date: 2016-09-05
 Authors@R: c(person("Claudia", "Vitolo", email = "cvitolodev@gmail.com", role = c("aut", "cre")))
 Maintainer: Claudia Vitolo <cvitolodev@gmail.com>
 URL: https://github.com/cvitolo/hddtools/
