### 2016-09-06 a0fd4b0
Fixed unit tests
