### 2016-09-06 345d9ab
Fixed typos
