### 2016-09-06 2c60319
Added try to tests
