+++ tests/testthat/test-TRMM.R
@@ -17,6 +17,6 @@ test_that("Test TRMM function", {
     bbox = bbox,
     outputfileLocation = "."), silent = TRUE)
 
-  expect_that(is.null(x), equals(TRUE))
+  expect_that(class(x) != "try-error", equals(TRUE))
 
 })
