### 2016-09-06 985b9fb
changed output of rmd to github_document
