+++ tests/testthat/test-TRMM.R
@@ -3,6 +3,7 @@ context("TRMM")
 test_that("Test TRMM function", {
 
   skip_on_cran()
+  skip_on_travis()
 
   # Define a bounding box
   bbox <- list(lonMin=-3.82, latMin=48,lonMax=-3.63, latMax=50)
@@ -17,6 +18,6 @@ test_that("Test TRMM function", {
     bbox = bbox,
     outputfileLocation = "."), silent = TRUE)
 
-  expect_that(class(x) != "try-error", equals(TRUE))
+  expect_that(class(x) == "try-error", equals(FALSE))
 
 })
