+++ README.md
@@ -1,29 +1,34 @@
 
+-   [hddtools: Hydrological Data Discovery Tools](#hddtools-hydrological-data-discovery-tools)
+    -   [Dependencies & Installation](#dependencies-installation)
+    -   [Data sources and Functions](#data-sources-and-functions)
+        -   [The Koppen Climate Classification map](#the-koppen-climate-classification-map)
+        -   [The Global Runoff Data Centre](#the-global-runoff-data-centre)
+        -   [NASA's Tropical Rainfall Measuring Mission (TRMM, only available for github version)](#nasas-tropical-rainfall-measuring-mission-trmm-only-available-for-github-version)
+        -   [Top-Down modelling Working Group (Data60UK and MOPEX)](#top-down-modelling-working-group-data60uk-and-mopex)
+        -   [SEPA river level data](#sepa-river-level-data)
+        -   [Met Office Hadley Centre Observation Data: HadUKP Data](#met-office-hadley-centre-observation-data-hadukp-data)
+    -   [T&C](#tc)
+    -   [Meta](#meta)
+
 <!-- Edit the README.Rmd only!!! The README.md is generated automatically from README.Rmd. -->
-
-
-
 hddtools: Hydrological Data Discovery Tools
 ===========================================
 
-[![DOI](https://zenodo.org/badge/9118/cvitolo/hddtools.svg)](https://zenodo.org/badge/latestdoi/9118/cvitolo/hddtools)
-[![Travis-CI Build Status](https://travis-ci.org/cvitolo/hddtools.svg?branch=master)](https://travis-ci.org/cvitolo/hddtools)
-[![AppVeyor Build Status](https://ci.appveyor.com/api/projects/status/github/cvitolo/hddtools?branch=master&svg=true)](https://ci.appveyor.com/project/cvitolo/hddtools)
-[![CRAN Status Badge](http://www.r-pkg.org/badges/version/hddtools)](https://cran.r-project.org/package=hddtools)
-[![CRAN Total Downloads](http://cranlogs.r-pkg.org/badges/grand-total/hddtools)](https://cran.r-project.org/package=hddtools)
-[![CRAN Monthly Downloads](http://cranlogs.r-pkg.org/badges/hddtools)](https://cran.r-project.org/package=hddtools)
+[![DOI](https://zenodo.org/badge/9118/cvitolo/hddtools.svg)](https://zenodo.org/badge/latestdoi/9118/cvitolo/hddtools) [![Travis-CI Build Status](https://travis-ci.org/cvitolo/hddtools.svg?branch=master)](https://travis-ci.org/cvitolo/hddtools) [![AppVeyor Build Status](https://ci.appveyor.com/api/projects/status/github/cvitolo/hddtools?branch=master&svg=true)](https://ci.appveyor.com/project/cvitolo/hddtools) [![CRAN Status Badge](http://www.r-pkg.org/badges/version/hddtools)](https://cran.r-project.org/package=hddtools) [![CRAN Total Downloads](http://cranlogs.r-pkg.org/badges/grand-total/hddtools)](https://cran.r-project.org/package=hddtools) [![CRAN Monthly Downloads](http://cranlogs.r-pkg.org/badges/hddtools)](https://cran.r-project.org/package=hddtools)
 
-**hddtools** stands for Hydrological Data Discovery Tools. This R package is an open source project designed to facilitate non-programmatic access to a variety of online open data sources relevant for hydrologists and, more in general, environmental scientists and practitioners. 
+**hddtools** stands for Hydrological Data Discovery Tools. This R package is an open source project designed to facilitate non-programmatic access to a variety of online open data sources relevant for hydrologists and, more in general, environmental scientists and practitioners.
 
[…]
