### 2016-09-06 e6a6e51
Cleanup the README
