### 2016-09-06 24487e6
Minor typos
