### 2016-09-06 #73 @cvitolo
Thanks @noamross! You are right, I forgot to remove the tests folder from the .gitignore file.

I have also amended the package based on the above suggestions.
I have two internal functions in the package: `bboxSpatialPolygon()` and `getContent()`. These are the only two functions for which I do not have unit tests.

Please let me know if there is any outstanding issue to resolve.
