### 2016-09-06 b23bcca
Added JOSS badge
