### 2016-09-06 #73 @mdsumner
@cvitolo just some unsolicited input here, I  noticed your mention of bboxSpatialPolygon:  I recently put "spex" package on CRAN which tries to fill this gap. An extra package depend is probably not welcome, but thought it might be helpful. I am constantly repeating the pattern of as(extent(x) , "SpatialPolygons") then putting on the CRS again etc. (Really, extent() should have CRS, and though the lower level sp::Spatial() does have it, we usually want an actual object and not just the abstraction).
