### 2016-09-06 #73 @noamross
@cvitolo I see that you submitted this package to JOSS separately: https://github.com/openjournals/joss-reviews/issues/56.  This is fine but the "automatic submission" option we have up top envisions a process where the package goes to JOSS after our review (so that they can approve based on our reviews, removing an additional review step). If this makes sense to you, perhaps you should pause the review process over at JOSS?
