### 2016-09-07 #73 @cvitolo
Of course, I just let JOSS know I have submitted to ropensci as well.
