### 2016-09-07 6211fbf
Deleted preparePackage.R
