### 2016-09-07 8513677
Added rgdal and tests to travis
