+++ tests/testthat/test-catalogueData60UK.R
@@ -3,7 +3,6 @@ context("Data60UK")
 test_that("Test catalogueData60UK function", {
 
   skip_on_cran()
-  skip_on_travis()
 
   # Retrieve the whole catalogue
   x <- try(catalogueData60UK(), silent = TRUE)
@@ -22,8 +21,6 @@ test_that("Test catalogueData60UK function", {
 })
 
 test_that("Test tsData60UK function", {
-
-  skip_on_cran()
 
   # Retrieve sample data
   x <- tsData60UK(hydroRefNumber = 39015)
