+++ NAMESPACE
@@ -11,6 +11,7 @@ export(tsData60UK)
 export(tsGRDC)
 export(tsMOPEX)
 export(tsSEPA)
+import(rgdal)
 importFrom(Hmisc,monthDays)
 importFrom(RCurl,getURL)
 importFrom(RCurl,url.exists)
