### 2016-09-09 2e2a895
Re-arranged badges
