### 2016-09-22 #73 @ropenscibot
@mdsumner @ledell - hey there, it's been 16 days, please get your review in by Sep 27, thanks :smiley_cat: (ropensci-bot)
