### 2016-10-05 #73 @ledell
@noamross @cvitolo Here is my review:

>##### Paper (for packages co-submitting to JOSS)
>
>The package contains a `paper.md` with:
>
>- [ ] **A short summary** describing the high-level functionality of the software
>- [ ] **Authors:**  A list of authors with their affiliations
>- [ ] **A statement of need** clearly staing problems the software is designed to solve and its target audience.
>- [ ] **References:** with DOIs for all those that have one (e.g. papers, datasets, software).

---

hddtools, which stands for Hydrological Data Discovery Tools, is an R interface to several online hydrology dataset repositories. The package is useful and relevant particularly to hydrologists, but also to environmental scientists and practitioners in general.

This package simplifies the following processes for the user: download of a metadata catalogue, selection of information needed, formal request for dataset(s), de-compression, conversion, manual filtering and parsing.

The package also can provide offline modes in addition to online modes, which makes it efficient to use if the user regularly works with the same datasets (no need to re-download on each session when data licenses allow for redistribution).

hddtools fills a gap in R data tools for hydrologists and environmental scientists alike and would be a good contribution to the rOpenSci package ecosystem.

#### Package Metadata

Does the package actually not work with R 3.0.1 or less, or was this just the version that was used in development and/or testing?

```
Suggests: testthat
```
I'd add the rgdal package to at least Suggests and maybe Depends since the `KGClimateClass()` function depends on it.

I'd possibly consider moving the Imports to Depends for the packages that provide primary utility to the hddtools package.

Are you familiar with permissive vs non-permissive licencing?  A majority of commercial businesses will stray away from GPL-licensed software, so you will have a potentially wider user base if you license with a permissive license such as MIT, BSD or Apache 2.0.  Just something to consider as you license your package, however if it's mostly academics who will use this package, then it probably doesn't matter too much.  If you want to keep it GPL'd, you may consider with GPL (>= 2) instead of GPL-3 because it will be more compatible with other packages.

In the description of the package, you state, "This R package is an open source project designed to facilitate non-programmatic access to a variety of online open data sources relevant for hydrologists and, more in general, environmental scientists and practitioners."

When I hear "non-programmatic access", I think of something like a GUI or website.  An R API would actually be the converse -- "programmatic access".  I would consider changing the wording here.

#### Code style

Try to avoid using a dot in variable or function names, as in `new.packages`.  I recommend the following edit to your README.md:

Change:

packs <- c('zoo', 'sp', 'RCurl', 'XML', 'rnrfa', 'Hmisc', 'raster',
           'stringr', 'devtools')
new.packages <- packs[!(packs %in% installed.packages()[,'Package'])]
if(length(new.packages)) install.packages(new.packages)
```

To this:

packs <- c('zoo', 'sp', 'RCurl', 'XML', 'rnrfa', 'Hmisc', 'raster',
           'stringr', 'devtools')
new_packs <- packs[!(packs %in% installed.packages()[,'Package'])]
if(length(new_packs)) install.packages(new_packs)
```

The convention for R code is to use spaces around all variable names and equals signs.  For example:

This:

bbox <- list(lonMin=-10,latMin=48,lonMax=5,latMax=62)
```

Instead of this:

bbox <- list(lonMin = -10, latMin = 48, lonMax = 5, latMax = 62)
```

##### Grammar/Spelling

[README.md](https://github.com/cvitolo/hddtools/blob/master/README.md):

- "formal request for" -> "a formal request for"
- "All those operation" -> "All those operations"
- LaTeX style quotes are not needed in Markdown: Update quotes ``Long-Term Mean Monthly Discharges'' (in [this](https://github.com/cvitolo/hddtools/#the-global-runoff-data-centre) section)
- You are missing the name of the function in this sentence, "The hddtools provides a function, called , to download "

I think it would be nice to have a few more hyperlinks to the data sources, for example, you could add a link to the GDRC or the Data60UK initiative.

[DESCRIPTION]()

- ": Many governmental bodies" -> "Many governmental bodies"
- "Especially when programmatic access to data resources is not allowed, downloading metadata catalogue, select the information needed, request datasets, de-compression, conversion, manual filtering and parsing can become rather tedious." -> "Especially when programmatic access to data resources is not allowed, downloading a metadata catalogue, selecting the information needed, requesting datasets, de-compression, conversion, manual filtering and parsing can become rather tedious."
- "The package facilitate non programmatic access to various online data sources" -> "The package facilitates programmatic access to various online data sources"
- "This package complements R's growing functionality in environmental web technologies to bridge the gap" -> "This package complements R's growing functionality in environmental web technologies by bridging the gap"
- "data consumers and it is designed to be the starting building block" -> "data consumers.  It is designed to be an initial building block"
- "Facilitates discovery and handling of hydrological data, non-programmatic access to catalogues and databases." -> "Facilitates discovery and handling of hydrological data via programmatic access to catalogues and databases."

There was a missing package when I tried to run the first example:

KGClimateClass(bbox,updatedBy='Peel')
Error in .rasterObjectFromFile(x, band = band, objecttype = "RasterLayer",  :
  Cannot create RasterLayer object from this file; perhaps you need to install rgdal first
```

To remedy this, either add this package to the required/dependent packages in the README.md Installation section, or add a line to install the package right before using the function that depends on it.  For example:

install.packages('rgdal')
KGClimateClass(bbox,updatedBy='Peel')
```

Found a possible bug:

> library(raster)
Loading required package: sp
> b <- brick('trmm_acc.tif')
> plot(b)
Error in graphics::par(old.par) :
  invalid value specified for graphical parameter "pin"
```

Is there any way you can add the size of the datasets (assuming they stay static), to the docmentation for the data downloading functions?  For example, this function:

x <- catalogueData60UK()
```

My plot for the following function looks quite different, I'm assuming that's because it's different data.  I'd add a comment about how the plots will look different to the user, but that the following are examples of what the plots may look like.

y <- tsSEPA(hydroRefNumber = c(234253, 234174, 234305))
plot(y[[1]])
```

Also, these plot all on the same plot (4 graphs on one plot).  Is that intentional?

> plot(y[[1]])
> plot(y[[2]])
> plot(y[[3]])
> x <- HadDAILY()
> plot(x$EWP)
```

#### R Documentation

Is `bboxSpatialPolygon` a user-level function or an internal function?  It's not available to the user, apparently...

```
> library(hddtools)
> bbox <- list(lonMin=-180,latMin=-50,lonMax=+180,latMax=+50)
> bbSP <- bboxSpatialPolygon(bbox)
Error: could not find function "bboxSpatialPolygon"
```

You can add linkable URLS using the `\url{}` command.  Your links in the R docs are not currently hyperlinked.

In the R docs for the catalogue functions, I'd save the output to a data.frame.  For example:

# Retrieve the whole catalogue
x <- catalogueGRDC()
```
The only one that currently has this is:

x <- catalogueSEPA()
```

The `getContent()` function is not exported correctly:

>   url <- "ftp://hydrology.nws.noaa.gov/pub/gcip/mopex/US_Data/Us_438_Daily"
>   getContent(url)
Error: could not find function "getContent"
```
The example for `tsData60UK()` is simply commented out instead of using the "Example Not Run" command.

Sometimes you specify the argument name in the example and sometimes you don't.  It might be best to fully specify the arguments in all the examples:

Specified:

x <- tsGRDC(stationID=1107700)
```

Not specified:

x <- tsMOPEX("14359000")
```

It would be good to note when a hydrometric reference number is a character string because some of the functions use numeric IDs and some use strings.

I'd change the title/name of "SEPAcatalogue" to "Data set: The SEPA Catalogue" to make sure that it's clear that it's a dataset. It might even be useful to call out any data sources as well using the term, "Data source: [name of catalogue]"
