### 2016-10-05 #73 @noamross
Thanks @ledell, for a thorough review! One quick thing - could you also look at the short paper (in `inst/paper`) for the JOSS section?
