### 2016-10-10 cvitolo/hddtools#2 @mdsumner
KGClimate-class problem
I see this problem, happy to help figure it out but for now I'll  just post the details.

``` R

library(hddtools)
# Define a bounding box
bbox <- list(lonMin=-10,latMin=48,lonMax=5,latMax=62)

# Define a temporal extent
timeExtent <- seq(as.Date('2012-01-01'), as.Date('2012-03-31'), by='months')

KGClimateClass(bbox,updatedBy='Peel')
# Error in readBin(tarfile, "raw", n = 3L) :
#   can only read from a binary connection
# In addition: Warning message:
#   In file(con, "rb") :
#   file("") only supports open = "w+" and open = "w+b": using the former

devtools::session_info()
# Session info -----------------------------------------------------------------------
#   setting  value
# version  R version 3.3.1 Patched (2016-10-08 r71472)
# system   x86_64, mingw32
# ui       RStudio (1.0.35)
# language (EN)
# collate  English_Australia.1252
# tz       Australia/Hobart
# date     2016-10-11
```
