### 2016-10-10 ropensci/hddtools#3 @mdsumner
problem with TRMM on Windows and Linux
Happy to explore this, for now just the details.

On Windows.

``` R
TRMM(timeExtent = timeExtent, bbox = bbox)
latMax of bbox is out of the maximum extent [-50,50], new latMax of bbox is modified as follows: latMax = 50
```

```

Beware this function was tested on a unix machine. If you are not using a unix machine your results could be wrong. Please report any problem to ...@gmail.com, thanks!
Error in .local(.Object, ...) :
  Unable to open ...3B43.*.
No error

Error in .rasterObjectFromFile(x, objecttype = "RasterBrick", ...) :
  Cannot create a RasterLayer object from this file.
```
