### 2016-10-11 cvitolo/r_hddtools#3 @mdsumner
On Debian.

``` R
 bbox <- list(lonMin=-10,latMin=48,lonMax=5,latMax=62)
>
> # Define a temporal extent
> timeExtent <- seq(as.Date('2012-01-01'), as.Date('2012-03-31'), by='months')
> TRMM(timeExtent = timeExtent, bbox = bbox)
latMax of bbox is out of the maximum extent [-50,50], new latMax of bbox is modified as follows: latMax = 50
Error in function (type, msg, asError = TRUE)  :
  Failed EPSV attempt, exiting
> devtools::session_info()
Session info -------------------------------------------------------------------
 setting  value
 version  R version 3.3.1 (2016-06-21)
 system   x86_64, linux-gnu
 ui       X11
 language (EN)
 collate  en_AU.UTF-8
 tz       Australia/Hobart
 date     2016-10-11

```
