### 2016-10-13 #73 @mdsumner
## Package Review

The package `hddtools` provides direct access to number of online data services for hydrological data. These data come in a variety of forms, and most include a catalogue of available data as an index for making further data requests. The package provides the data in easily useable forms, leveraging existing tools in R for mapping, spatial data handling, and time series handling and plotting.

I have made a number of detailed suggestions below, and I'm happy to follow up and help with any of those.

Finally, apologies for the delay in this review.

#### Documentation

- [x ] **A statement of need** clearly stating problems the software is designed to solve and its target audience in README
- [x ] **Installation instructions:** for the development version of package and any non-standard dependencies in README
- [x ] **Vignette(s)** demonstrating major functionality that runs successfully locally
- [x ] **Function Documentation:** for all exported functions in R help
- [x ] **Examples** for all exported functions in R Help that run successfully locally
- [x ] **Community guidelines** including contribution guidelines in the README or CONTRIBUTING, and `URL`, `Maintainer` and `BugReports` fields in DESCRIPTION

I installed into a fresh R 3.3.1 on Windows, using the Github dev instructions. The devtools github install added rgdal during the process.

I had some issues with some functions, which I submitted as issues - but it may well be a problem specific to a local machine I have - it works on different machine.

https://github.com/cvitolo/hddtools/issues/2

https://github.com/cvitolo/hddtools/issues/3

>##### Paper (for packages co-submitting to JOSS)
>
>The package contains a `paper.md` with:
>
>- [x ] **A short summary** describing the high-level functionality of the software
>- [ ] **Authors:**  A list of authors with their affiliations
>- [x ] **A statement of need** clearly staing problems the software is designed to solve and its target audience.
>- [x ] **References:** with DOIs for all those that have one (e.g. papers, datasets, software).

There's no list of authors here, but I haven't explored what is required in detail either.

#### Functionality

- [x ] **Installation:** Installation succeeds as documented.
- [x ] **Functionality:** Any functional claims of the software been confirmed.
- [x ] **Performance:** Any performance claims of the software been confirmed.
- [x ] **Packaging guidelines**: The package conforms to the rOpenSci packaging guidelines

I had some quirks with TRMM function download not working correctly on specific machines, I will try to follow it up: https://github.com/cvitolo/hddtools/issues/3 (because now it's working for me).

Code coverage is curretly at 55% so I recommend putting in more tests to increse this.

#### Final approval (post-review)

Estimated hours spent reviewing:

---

### Review Comments

I'd like to see the examples explore the catalogue data so there's more clarity about what data are available and how to use the filters.

Some catalogue help pages have a full listing of their contents (e.g. SEPA ) but others do not.

### Suggestions

* It would be handy to see links to all the providers of data (GRDC, MOPEX, Hadley etc) in the hddtools-package doc page
* use `raster::extent()` rather than a raw list,  could migrate to this default by providing an extent method for your raw lists
* note that `bbox` is a function in `sp`, so you might want a different name
* clarify `timeExtent` is really an axis of steps, not a range/extent - though I think it is only used as a range (??) so range(timeExtent) would do, the seq(,by=) part is not relevant I think
* include checks for sense in bounding boxes, I inadvertently tried `bb <- list(lonMin=-100,latMin=43,lonMax=-50,latMax=42)` which doesn't complain but will never return results (this would be provided automatically by using raster::extent)
* README example  KGClimateClass: https://github.com/cvitolo/hddtools/issues/2
* examples use `x` for different calls to `catalogueGRDC`, please name these objects differently for interactive comparison, and do this throughout examples
* clarify in comments that filter is bbox plus "keep only those where "statistics" value is 1""
* leaflet installation should be suggested before `library(leaflet)`
* links to GRDC and KGClimate sites, I was looking for descriptions of the columns in catalogueGRDC() and/or a link to the GRDC resources that describe this
* README: "and, more in general, environmental scientists and practitioners." remove the "more" ?
* "``Long-Term Mean Monthly Discharges''" in README has funky quotations
* I suggest not using generic (and repeated) variable names like "x" and "y", use specific names for the data - so an object is standalone from examples
* use links to URLs in help pages
* getContent is documented but not exported

* provide example of how to discover possible values for `stationID` for `tsGRDC`, and/or link to `catalogueGRDC` function in help for `stationID` argument

E.g.
```R
bb <- list(lonMin=-100,latMin=42,lonMax=-50,latMax=43)

cG <- catalogueGRDC(bbox = bb)

tsg <- tsGRDC(stationID = cG$grdc_no[2])

```

*  always create the bbox and timeExtent explicitly for each example, so their scope is narrow and easily runnable.
* show that y can be plotted directly here

```
hydroRefNumber <- as.numeric(as.character(x$id[1])) # e.g. 55012
y <- tsData60UK(hydroRefNumber, plotOption=FALSE)
plot(y)

```
* I ran this twice, expected the download to be cached? It downloads again

```
y <- tsGRDC(stationID='6122300', plotOption=TRUE)
trying URL 'ftp://ftp.bafg.de/pub/REFERATE/GRDC/ltdata/europe.zip'
```

* this example doesn't link the name of the file to the read, trmm_acc.tif could be an output value of TRMM?

```
bbox <-
TRMM(timeExtent = timeExtent, bbox = bbox)
library(raster)
b <- brick('trmm_acc.tif')
plot(b)
```

* doc in TRMM about value is not complete "Data is loaded as rasterbrick, then converted to a multilayer Geotiff that can" also, it's really the opposite - "multilayer GeoTIFF is downloaded, loaded as a RasterBrick"

Suggest a better workflow is this, but it could go further and build the RasterBrick and return it. The raster object will record that it's linked to a file, so that seems reasonable.

```
trmmfile <- TRMM(timeExtent = timeExtent, bbox = bbox)
library(raster)
b <- brick(trmmfile)
```

* I get this problem, but on one machine it worked fine - I'll try to follow up

```
 TRMM(timeExtent = timeExtent, bbox = bbox)
latMax of bbox is out of the maximum extent [-50,50], new latMax of bbox is modified as follows: latMax = 50
Beware this function was tested on a unix machine. If you are not using a unix machine your results could be wrong. Please report any problem to cvitolodev@gmail.com, thanks!
Error in .local(.Object, ...) :
  Unable to open C:\Users\michae_sum\Documents\projects\hddtools\3B43.*.
No error

 Show Traceback

 Rerun with Debug
 Error in .rasterObjectFromFile(x, objecttype = "RasterBrick", ...) :
  Cannot create a RasterLayer object from this file.
  ```
