### 2016-10-15 77b34f7
Appveyor testing R 3.0.1
