+++ README.Rmd
@@ -233,14 +233,10 @@ For example, here is how to plot the time series for England and Wales:
 plot(x$EWP)
 ```
 
-## T&C
-
-Please note that this project is released with a [Contributor Code of Conduct](CONDUCT.md). By participating in this project you agree to abide by its terms.
-
-This package and functions herein are part of an experimental open-source project. They are provided as is, without any guarantee.
-
 ## Meta
 
+* This package and functions herein are part of an experimental open-source project. They are provided as is, without any guarantee.
+* Please note that this project is released with a [Contributor Code of Conduct](CONDUCT.md). By participating in this project you agree to abide by its terms.
 * Please [report any issues or bugs](https://github.com/cvitolo/hddtools/issues).
 * License: [GPL-3](https://opensource.org/licenses/GPL-3.0)
 * Get citation information for `hddtools` in R doing `citation(package = 'hddtools')`
