### 2016-10-15 26d1270
Appveyor testing R pached, oldrel and R 3.0.1
