+++ appveyor.yml
@@ -20,8 +20,11 @@ environment:
 
   matrix:
   - R_VERSION: stable
-  - R_VERSION: 3.0.1
+  - R_VERSION: patched
+  - R_VERSION: oldrel
     RTOOLS_VERSION: 34
+  - R_VERSION: 3.0.1
+    RTOOLS_VERSION: 33
 
 build_script:
   - travis-tool.sh install_deps
