### 2016-10-15 db87795
Appveyor tesing R 3.0.1. and 2.5.0
