+++ appveyor.yml
@@ -22,9 +22,8 @@ environment:
   - R_VERSION: stable
   - R_VERSION: patched
   - R_VERSION: oldrel
-    RTOOLS_VERSION: 34
   - R_VERSION: 3.0.1
-    RTOOLS_VERSION: 33
+  - R_VERSION: 2.5.0
 
 build_script:
   - travis-tool.sh install_deps
