### 2016-10-15 8b30bff
Appveyor testing R 3.1, 3.2 and 3.0.2
