### 2016-10-16 ff29280
Appveyor test R 3.0.1 with rtools 33
