### 2016-10-16 8431a8e
Appveyor Fixed R-Rtools version compatibility
