### 2016-10-16 9448af2
Appveyor test  R_VERSION: 3.0 and RTOOLS_VERSION: 30
