### 2016-10-16 bd9ffdf
Appveyor R 3.1 Rtools 31
