+++ appveyor.yml
@@ -22,8 +22,8 @@ environment:
   - R_VERSION: stable
   - R_VERSION: patched
   - R_VERSION: oldrel
-  - R_VERSION: 3.0
-    RTOOLS_VERSION: 30
+  - R_VERSION: 3.1
+    RTOOLS_VERSION: 31
 
 build_script:
   - travis-tool.sh install_deps
