### 2016-10-16 b808315
Appveyor removed tests with old r-rtools versions
