### 2016-10-16 e17689f
Simplified README
