+++ README.Rmd
@@ -1,7 +1,5 @@
 ---
-output: 
-  github_document:
-    toc: true
+output: github_document
 ---
 
 <!-- Edit the README.Rmd only!!! The README.md is generated automatically from README.Rmd. -->
@@ -31,9 +29,9 @@ hddtools: Hydrological Data Discovery Tools
 [![AppVeyor Build Status](https://ci.appveyor.com/api/projects/status/github/cvitolo/hddtools?branch=master&svg=true)](https://ci.appveyor.com/project/cvitolo/hddtools)
 [![codecov.io](https://codecov.io/github/cvitolo/hddtools/coverage.svg?branch=master)](https://codecov.io/github/cvitolo/hddtools?branch=master)
 
-**hddtools** stands for Hydrological Data Discovery Tools. This R package is an open source project designed to facilitate non-programmatic access to a variety of online open data sources relevant for hydrologists and, more in general, environmental scientists and practitioners. 
+**hddtools** stands for Hydrological Data Discovery Tools. This R package is an open source project designed to facilitate access to a variety of online open data sources relevant for hydrologists and, more in general, environmental scientists and practitioners. 
 
-This typically implies the download of a metadata catalogue, selection of information needed, formal request for dataset(s), de-compression, conversion, manual filtering and parsing. All those operation are made more efficient by re-usable functions. 
+This typically implies the download of a metadata catalogue, selection of information needed, a formal request for dataset(s), de-compression, conversion, manual filtering and parsing. All those operations are made more efficient by re-usable functions. 
 
 Depending on the data license, functions can provide offline and/or online modes. When redistribution is allowed, for instance, a copy of the dataset is cached within the package and updated twice a year. This is the fastest option and also allows offline use of package's functions. When re-distribution is not allowed, only online mode is provided.
 
@@ -41,197 +39,47 @@ Depending on the data license, functions can provide offline and/or online modes
 The hddtools package depends on other CRAN packages. Check for missing dependencies and install them using the commands below:
 
 ```{r}
-packs <- c('zoo', 'sp', 'RCurl', 'XML', 'rnrfa', 'Hmisc', 'raster', 
-           'stringr', 'devtools')
-new.packages <- packs[!(packs %in% installed.packages()[,'Package'])]
-if(length(new.packages)) install.packages(new.packages)
+packs <- c("zoo", "sp", "RCurl", "XML", "rnrfa", "Hmisc", "raster", 
+           "stringr", "devtools")
+new_packages <- packs[!(packs %in% installed.packages()[,"Package"])]
+if(length(new_packages)) install.packages(new_packages)
 ```
 
 Get the released version from CRAN:
 
 ```{r}
-install.packages('hddtools')
+install.packages("hddtools")
 ```
 
 Or the development version from github using [devtools](https://github.com/hadley/devtools):
 
 ```{r}
-# install.packages("devtools")
-devtools::install_github('cvitolo/hddtools')
+devtools::install_github("cvitolo/hddtools")
 ```
 
 Load the hddtools package:
 
 ```{r, eval = TRUE}
-library('hddtools')
+library("hddtools")
 ```
 
 ## Data sources and Functions
 
-The functions provided can filter, slice and dice 2 and 3 dimensional information. 
+The package contains functions to interact with the data providers listed below. For examples of the various functionalities see the [vignette](vignettes/hddtools_vignette.Rmd).
 
-```{r, eval = TRUE}
-# Define a bounding box
[…]
