+++ README.md
@@ -1,15 +1,4 @@
 
--   [hddtools: Hydrological Data Discovery Tools](#hddtools-hydrological-data-discovery-tools)
-    -   [Dependencies & Installation](#dependencies-installation)
-    -   [Data sources and Functions](#data-sources-and-functions)
-        -   [The Koppen Climate Classification map](#the-koppen-climate-classification-map)
-        -   [The Global Runoff Data Centre](#the-global-runoff-data-centre)
-        -   [NASA's Tropical Rainfall Measuring Mission (TRMM, only available for github version)](#nasas-tropical-rainfall-measuring-mission-trmm-only-available-for-github-version)
-        -   [Top-Down modelling Working Group (Data60UK and MOPEX)](#top-down-modelling-working-group-data60uk-and-mopex)
-        -   [SEPA river level data](#sepa-river-level-data)
-        -   [Met Office Hadley Centre Observation Data: HadUKP Data](#met-office-hadley-centre-observation-data-hadukp-data)
-    -   [Meta](#meta)
-
 <!-- Edit the README.Rmd only!!! The README.md is generated automatically from README.Rmd. -->
 hddtools: Hydrological Data Discovery Tools
 ===========================================
@@ -20,9 +9,9 @@ hddtools: Hydrological Data Discovery Tools
 
 [![Travis-CI Build Status](https://travis-ci.org/cvitolo/hddtools.svg?branch=master)](https://travis-ci.org/cvitolo/hddtools) [![AppVeyor Build Status](https://ci.appveyor.com/api/projects/status/github/cvitolo/hddtools?branch=master&svg=true)](https://ci.appveyor.com/project/cvitolo/hddtools) [![codecov.io](https://codecov.io/github/cvitolo/hddtools/coverage.svg?branch=master)](https://codecov.io/github/cvitolo/hddtools?branch=master)
 
-**hddtools** stands for Hydrological Data Discovery Tools. This R package is an open source project designed to facilitate non-programmatic access to a variety of online open data sources relevant for hydrologists and, more in general, environmental scientists and practitioners.
+**hddtools** stands for Hydrological Data Discovery Tools. This R package is an open source project designed to facilitate access to a variety of online open data sources relevant for hydrologists and, more in general, environmental scientists and practitioners.
 
-This typically implies the download of a metadata catalogue, selection of information needed, formal request for dataset(s), de-compression, conversion, manual filtering and parsing. All those operation are made more efficient by re-usable functions.
+This typically implies the download of a metadata catalogue, selection of information needed, a formal request for dataset(s), de-compression, conversion, manual filtering and parsing. All those operations are made more efficient by re-usable functions.
 
 Depending on the data license, functions can provide offline and/or online modes. When redistribution is allowed, for instance, a copy of the dataset is cached within the package and updated twice a year. This is the fastest option and also allows offline use of package's functions. When re-distribution is not allowed, only online mode is provided.
 
@@ -32,235 +21,48 @@ Dependencies & Installation
 The hddtools package depends on other CRAN packages. Check for missing dependencies and install them using the commands below:
 
 ``` r
-packs <- c('zoo', 'sp', 'RCurl', 'XML', 'rnrfa', 'Hmisc', 'raster', 
-           'stringr', 'devtools')
-new.packages <- packs[!(packs %in% installed.packages()[,'Package'])]
-if(length(new.packages)) install.packages(new.packages)
[…]
