### 2016-10-16 2510e40
Removed the expression 'non-programmatic' from the documention. Avoid using a dot in variable or function names. Use spaces around all variable names and equals signs.
