+++ inst/paper/paper.md
@@ -7,14 +7,14 @@ tags:
 - hydrology
 - R
 authors:
-- affiliation: Brunel University London
+- affiliation: European Centre for Medium-range Weather Forecasts
   name: Claudia Vitolo
   orcid: 0000-0002-4252-1176
 ---
 
 # Summary
 
-The hddtools [@hddtools-archive] (**h**ydrological **d**ata **d**iscovery **tools**) is an R package [@R-base] designed to facilitate non-programmatic access to a variety of online open data sources relevant for hydrologists and, more in general, environmental scientists and practitioners. This typically implies the download of a metadata catalogue, selection of information needed, formal request for dataset(s), de-compression, conversion, manual filtering and parsing. All those operation are made more efficient by re-usable functions. 
+The hddtools [@hddtools-archive] (**h**ydrological **d**ata **d**iscovery **tools**) is an R package [@R-base] designed to facilitate access to a variety of online open data sources relevant for hydrologists and, more in general, environmental scientists and practitioners. This typically implies the download of a metadata catalogue, selection of information needed, formal request for dataset(s), de-compression, conversion, manual filtering and parsing. All those operation are made more efficient by re-usable functions. 
 
 Depending on the data license, functions can provide offline and/or online modes. When redistribution is allowed, for instance, a copy of the dataset is cached within the package and updated twice a year. This is the fastest option and also allows offline use of package's functions. When re-distribution is not allowed, only online mode is provided.
 
