+++ R/TRMM.R
@@ -24,8 +24,8 @@
 #' @examples
 #' \dontrun{
 #'   # Define a bounding box
-#'   bbox <- list(lonMin=-3.82, latMin=48,lonMax=-3.63, latMax=50)
-#'   twindow <- seq(as.Date("2012-01-01"), as.Date("2012-01-31"), by="months")
+#'   bbox <- list(lonMin = -3.82, latMin = 48, lonMax = -3.63, latMax = 50)
+#'   twindow <- seq(as.Date("2012-01-01"), as.Date("2012-01-31"), by = "months")
 #'
 #'   TRMM(inputLocation="ftp://disc2.nascom.nasa.gov/data/TRMM/Gridded/",
 #'        product = "3B43",
@@ -58,7 +58,7 @@ TRMM <- function(inputLocation="ftp://disc2.nascom.nasa.gov/data/TRMM/Gridded/",
   # Check bounding box extent is within original raster extent
   if (is.null(bbox)){
     # Use TRMM max extent
-    bbox <- list(lonMin=-180,latMin=-50,lonMax=+180,latMax=+50)
+    bbox <- list(lonMin = -180, latMin = -50, lonMax = +180, latMax = +50)
   }
   if (bbox$lonMin < -180 | bbox$lonMin > 180) {
     bbox$lonMin <- -180
@@ -87,10 +87,10 @@ TRMM <- function(inputLocation="ftp://disc2.nascom.nasa.gov/data/TRMM/Gridded/",
   bbSP <- bboxSpatialPolygon(bbox)
 
   # Check time extent
-  years <- unique(format(timeExtent,"%Y"))
-  months <- format(timeExtent,"%m")
+  years <- unique(format(timeExtent, "%Y"))
+  months <- format(timeExtent, "%m")
 
-  if ( any(years=="NULL",months=="NULL") ){
+  if ( any(years == "NULL", months == "NULL") ){
 
     message(paste("This function cannot be executed because your timeExtent",
                   "is not in the correct format.",
@@ -103,7 +103,7 @@ TRMM <- function(inputLocation="ftp://disc2.nascom.nasa.gov/data/TRMM/Gridded/",
     for (myYear in years){
 
       myURL <- paste(inputLocation,
-                     product, "_V", version, "/", myYear, "/", sep="")
+                     product, "_V", version, "/", myYear, "/", sep = "")
 
       filenames <- getURL(myURL,
                           ftp.use.epsv = FALSE,
@@ -114,10 +114,10 @@ TRMM <- function(inputLocation="ftp://disc2.nascom.nasa.gov/data/TRMM/Gridded/",
       # e.g. only certain months.
       # "*precipitation.accum" means monthly accumulated rainfall here.
       for (myMonth in months){
-        filePaths <- paste(myURL,product,".",substr(myYear,3,4),myMonth,"01",
-                           ".",version,".",type,sep="")
+        filePaths <- paste(myURL, product, ".", substr(myYear, 3, 4),
+                           myMonth, "01", ".", version, ".", type, sep = "")
 
-        selectedfilePaths <- c(selectedfilePaths,filePaths)
+        selectedfilePaths <- c(selectedfilePaths, filePaths)
       }
 
     }
@@ -125,7 +125,7 @@ TRMM <- function(inputLocation="ftp://disc2.nascom.nasa.gov/data/TRMM/Gridded/",
     # download files
     # Thanks to Fabien Wagner this also works for windows' users!
     mapply(download.file, selectedfilePaths,
-           basename(selectedfilePaths), method="wget")
+           basename(selectedfilePaths), method = "wget")
 
     # Now I create a virtual file as the downloaded TRMM data come as binaries.
[…]
