### 2016-10-16 ef246a1
Amend various sentences in hddtools_package.R
