### 2016-10-16 8d51a03
links in the R docs are now hyperlinked
