+++ man/bboxSpatialPolygon.Rd
@@ -29,6 +29,6 @@ Bounding box is first created (in lat/lon) then projected if specified
 
 }
 \references{
-http://gis.stackexchange.com/questions/46954/clip-spatial-object-to-bounding-box-in-r
+\url{http://gis.stackexchange.com/questions/46954/clip-spatial-object-to-bounding-box-in-r}
 }
 
