+++ man/KGClimateClass.Rd
@@ -33,6 +33,6 @@ Given a bounding box, the function identifies the overlapping climate zones.
 Claudia Vitolo
 }
 \references{
-Kottek et al. (2006): http://koeppen-geiger.vu-wien.ac.at/. Peel et al. (2007): http://people.eng.unimelb.edu.au/mpeel/koppen.html.
+Kottek et al. (2006): \url{http://koeppen-geiger.vu-wien.ac.at/}. Peel et al. (2007): \url{http://people.eng.unimelb.edu.au/mpeel/koppen.html}.
 }
 

[…]
