### 2016-10-16 cb103a8
The arguments in all the examples are now fully specified.
