+++ R/bboxSpatialPolygon.R
@@ -14,7 +14,7 @@
 #' @examples
 #' \dontrun{
 #'   bbox <- list(lonMin = -180, latMin = -50, lonMax = +180, latMax = +50)
-#'   bbSP <- bboxSpatialPolygon(bbox)
+#'   bbSP <- bboxSpatialPolygon(bbox = bbox)
 #' }
 #'
 
