+++ man/KGClimateClass.Rd
@@ -25,7 +25,7 @@ Given a bounding box, the function identifies the overlapping climate zones.
   bbox <- list(lonMin = -3.82, latMin = 52.41,
                lonMax = -3.63, latMax = 52.52)
   # Get climate classes
-  KGClimateClass(bbox)
+  KGClimateClass(bbox = bbox)
 }
 
 }
