### 2016-10-16 89859e7
The hydrometric reference number is now always a string.
