+++ R/Data60UK.R
@@ -106,7 +106,7 @@ catalogueData60UK <- function(bbox = NULL, columnName = NULL,
 #'
 #' @description This function extract the dataset containing daily rainfall and streamflow discharge at one of the Data60UK locations.
 #'
-#' @param hydroRefNumber hydrometric reference number
+#' @param hydroRefNumber hydrometric reference number (string)
 #' @param plotOption boolean to define whether to plot the results. By default this is set to TRUE.
 #' @param timeExtent is a vector of dates and times for which the data should be retrieved
 #'
