### 2016-10-16 9b71bbb
The title of SEPAcatalogue is now Data set: The SEPA Catalogue
