### 2016-10-16 #73 @cvitolo
Thanks for the comments @ledell @mdsumner! I'll try and address them all in the next couple of weeks.
