### 2016-11-01 f16d54a
Addressed all the comments from first reviewer
