+++ tests/testthat/test-GRDC.R
@@ -12,7 +12,7 @@ test_that("Test catalogueGRDC function", {
   bbox <- list(lonMin=-3.82,latMin=52.41,lonMax=-3.63,latMax=52.52)
   # Filter the catalogue
   x <- catalogueGRDC(bbox)
-  expect_that(all(dim(x) == c(2, 44)), equals(TRUE))
+  expect_that(all(dim(x) == c(2, 46)), equals(TRUE))
 
 })
 
@@ -21,7 +21,7 @@ test_that("Test tsGRDC function", {
   skip_on_cran()
 
   # Retrieve sample data
-  x <- tsGRDC(stationID=1107700)
+  x <- tsGRDC(stationID = 1107700)
   expect_that("list" %in% class(x), equals(TRUE))
   expect_that(all(names(x) == c("mddPerYear", "mddAllPeriod", "mddPerMonth")),
               equals(TRUE))

[…]
