### 2016-11-04 #73 @sckott
@mdsumner do you remember your estimated hours spent reviewing?
