### 2016-11-05 #73 @mdsumner
4.5 hours

On Sat, Nov 5, 2016, 03:39 Scott Chamberlain <notifications@github.com>
wrote:

> @mdsumner <https://github.com/mdsumner> do you remember your estimated
> hours spent reviewing?
>
