### 2016-12-18 cvitolo/hddtools#4 @cvitolo
Error in graphics::par(old.par)
This issue is part of this [ropensci review](https://github.com/ropensci/onboarding/issues/73)

@ledell said:

Found a possible bug:
library(raster)
Loading required package: sp
b <- brick('trmm_acc.tif')
plot(b)
Error in graphics::par(old.par) :
invalid value specified for graphical parameter "pin"
