### 2016-12-18 cvitolo/r_hddtools#4 @cvitolo
@ledell I cannot reproduce your error here.

Have you generated the file trmm_acc.tiff using the `TRMM()` function before running the `brick()` command?

@mdsumner suggested to change slightly the workflow for the TRMM function.
Please install the latest version of hddtools from github and test the TRMM function using the example code below. Please let me know if the problem persists, thanks!

```
# Define a bounding box
areaBox <- raster::extent(-10, 5, 48, 62)
# Define a temporal extent
twindow <- seq(as.Date("2012-01-01"), as.Date("2012-03-31"), by = "months")
# Retreive mean monthly precipitations from 3B43_V7 (based on a bounding box and time extent)
TRMMfile <- TRMM(twindow = twindow, areaBox = areaBox)
library(raster)
plot(TRMMfile)
```
