### 2016-12-18 ed04dce
Addressed all comments from first reviewer
