+++ R/KGClimateClass.R
@@ -5,7 +5,7 @@
 #' @description Given a bounding box, the function identifies the overlapping climate zones.
 #'
 #' @param bbox bounding box, a list made of 4 elements: minimum longitude (lonMin), minimum latitude (latMin), maximum longitude (lonMax), maximum latitude (latMax)
-#' @param updatedBy this can either be Kottek
+#' @param updatedBy this can either be "Kottek" or "Peel"
 #' @param verbose if TRUE more info are printed on the screen
 #'
 #' @references Kottek et al. (2006): \url{http://koeppen-geiger.vu-wien.ac.at/}. Peel et al. (2007): \url{http://people.eng.unimelb.edu.au/mpeel/koppen.html}.
