### 2016-12-18 f248342
Added catalogue information to all help pages
