+++ R/Data60UK.R
@@ -19,8 +19,7 @@
 #'   x <- catalogueData60UK()
 #'
 #'   # Define a bounding box
-#'   bbox <- list(lonMin = -3.82, latMin = 52.41,
-#'                lonMax = -3.63, latMax = 52.52)
+#'   bbox <- list(lonMin = -4, latMin = 52, lonMax = -2, latMax = 53)
 #'
 #'   # Filter the catalogue
 #'   x <- catalogueData60UK(bbox)
