+++ tests/testthat/test-GRDC.R
@@ -9,7 +9,7 @@ test_that("Test catalogueGRDC function", {
   expect_that("data.frame" %in% class(x), equals(TRUE))
 
   # Define a bounding box
-  bbox <- list(lonMin=-3.82,latMin=52.41,lonMax=-3.63,latMax=52.52)
+  bbox <- list(lonMin=-3.82, latMin=52.41, lonMax=-3.63, latMax=52.52)
   # Filter the catalogue
   x <- catalogueGRDC(bbox)
   expect_that(all(dim(x) == c(2, 46)), equals(TRUE))
