+++ R/SEPA.R
@@ -4,12 +4,11 @@
 #'
 #' @description This function provides an unofficial SEPA database catalogue of river level data (available from http://pennine.ddns.me.uk/riverlevels/ConciseList.html) containing info for 1752 stations. Some are NRFA stations.
 #'
-#' @param bbox bounding box, a list made of 4 elements: minimum longitude (lonMin), minimum latitude (latMin), maximum longitude (lonMax), maximum latitude (latMax)
-#' @param metadataColumn name of the column to filter
-#' @param entryValue value to look for in the column named metadataColumn
+#' @param columnName name of the column to filter
+#' @param columnValue value to look for in the column named columnName
 #' @param verbose if TRUE it returns whether the data is coming from live or cached data sources
 #'
-#' @return This function returns a data frame made of 830 rows (stations) and 8 columns: "idNRFA","aspxpage", "stationId", "River", "Location", "GridRef", "Operator" and "CatchmentArea(km2)". Column idNRFA shows the National River Flow Archive station id. Column "aspxpage" returns the Environment Agency gauges id. The column "stationId" is the id number used by SEPA. Use these id numbers to retrieve the time series of water levels.
+#' @return This function returns a data frame made of a maximum of 830 rows (stations) and 8 columns: "idNRFA","aspxpage", "stationId", "River", "Location", "GridRef", "Operator" and "CatchmentArea(km2)". Column idNRFA shows the National River Flow Archive station id. Column "aspxpage" returns the Environment Agency gauges id. The column "stationId" is the id number used by SEPA. Use these id numbers to retrieve the time series of water levels.
 #'
 #' @export
 #'
@@ -17,11 +16,17 @@
 #' \dontrun{
 #'   # Retrieve the whole catalogue
 #'   x <- catalogueSEPA()
+#'
+#'   # Get only catchments with area above 5000 Km2
+#'   x <- catalogueSEPA(columnName = "CatchmentAreaKm2",
+#'                      columnValue = ">= 5000")
+#'
+#'   # Get only catchments within river Avon
+#'   x <- catalogueSEPA(columnName = "River", columnValue = "Avon")
 #' }
 #'
 
-catalogueSEPA <- function(bbox = NULL,
-                          metadataColumn = NULL, entryValue = NULL,
+catalogueSEPA <- function(columnName = NULL, columnValue = NULL,
                           verbose = FALSE){
 
   theurl <- "http://pennine.ddns.me.uk/riverlevels/ConciseList.html"
@@ -35,7 +40,7 @@ catalogueSEPA <- function(bbox = NULL,
     row.names(SEPAcatalogue) <- NULL
     names(SEPAcatalogue) <- c("idNRFA", "aspxpage", "stationId", "River",
                               "Location",
-                              "GridRef", "Operator", "CatchmentArea(km2)")
+                              "GridRef", "Operator", "CatchmentAreaKm2")
   }else{
     if (verbose == TRUE) {
       message(paste("The connection with the live web data source failed.",
@@ -44,7 +49,43 @@ catalogueSEPA <- function(bbox = NULL,
     load(system.file("data/SEPAcatalogue.rda", package = "hddtools"))
   }
 
-  return(SEPAcatalogue)
+  SEPAcatalogue[] <- lapply(SEPAcatalogue, as.character)
+  SEPAcatalogue$CatchmentAreaKm2 <- as.numeric(SEPAcatalogue$CatchmentAreaKm2)
+
+  if ( !is.null(columnName) & !is.null(columnValue) ){
+
+    if (tolower(columnName) %in% tolower(names(SEPAcatalogue))){
+
[…]

[…]
