### 2016-12-20 bb2e4d3
Renamed bbox to areaBox and updated documentation
