+++ R/bboxSpatialPolygon.R
@@ -1,8 +1,8 @@
 #' Convert a bounding box to a SpatialPolygons object
 #' Bounding box is first created (in lat/lon) then projected if specified
 #'
-#' @param bbox Bounding box: a 2x2 numerical matrix of lat/lon coordinates
-#' @param proj4stringFrom Projection string for the current bbox coordinates (defaults to lat/lon, WGS84)
+#' @param boundingbox Bounding box: a 2x2 numerical matrix of lat/lon coordinates
+#' @param proj4stringFrom Projection string for the current boundingbox coordinates (defaults to lat/lon, WGS84)
 #' @param proj4stringTo Projection string, or NULL to not project
 #'
 #' @return A SpatialPolygons object of the bounding box
@@ -13,12 +13,12 @@
 #'
 #' @examples
 #' \dontrun{
-#'   bbox <- list(lonMin = -180, latMin = -50, lonMax = +180, latMax = +50)
-#'   bbSP <- bboxSpatialPolygon(bbox = bbox)
+#'   boundingbox <- raster::extent(c(-180, +180, -50, +50))
+#'   bbSP <- bboxSpatialPolygon(boundingbox = boundingbox)
 #' }
 #'
 
-bboxSpatialPolygon <- function(bbox,
+bboxSpatialPolygon <- function(boundingbox,
                                proj4stringFrom = NULL,
                                proj4stringTo = NULL) {
 
@@ -35,13 +35,13 @@ bboxSpatialPolygon <- function(bbox,
 
   }
 
-  bb <- matrix(as.numeric(c(bbox$lonMin, bbox$latMin,
-                            bbox$lonMax, bbox$latMax)),
+  bb <- matrix(as.numeric(c(boundingbox@xmin, boundingbox@ymin,
+                            boundingbox@xmax, boundingbox@ymax)),
                nrow=2)
   rownames(bb) <- c("lon", "lat")
   colnames(bb) <- c("min", "max")
 
-  # Create unprojected bbox as spatial object
+  # Create unprojected boundingbox as spatial object
   # clockwise, 5 points to close it
   bboxMat <- rbind( c(bb["lon", "min"], bb["lat", "min"]),
                     c(bb["lon", "min"], bb["lat", "max"]),
