+++ tests/testthat/test-catalogueData60UK.R
@@ -1,18 +1,27 @@
 context("Data60UK")
 
-test_that("Test catalogueData60UK function", {
-
-  skip_on_cran()
+test_that("Test full catalogueData60UK", {
 
   # Retrieve the whole catalogue
   x <- try(catalogueData60UK(), silent = TRUE)
+
   expect_that(class(x) != "try-error", equals(TRUE))
+  expect_that(all(dim(x) == c(61, 6)), equals(TRUE))
+
+})
+
+test_that("Test catalogueData60UK bounding box", {
 
   # Define a bounding box
-  bbox <- list(lonMin=-4,latMin=52,lonMax=-3,latMax=53)
+  areaBox <- raster::extent(c(-4, -2, +52, +53))
   # Filter the catalogue
-  x <- catalogueData60UK(bbox)
-  expect_that(all(dim(x) == c(2, 6)), equals(TRUE))
+  x <- catalogueData60UK(areaBox)
+
+  expect_that(all(dim(x) == c(6, 6)), equals(TRUE))
+
+})
+
+test_that("Test catalogueData60UK id", {
 
   # Filter the catalogue based on id
   x <- catalogueData60UK(columnName="id",columnValue="62001")
