+++ R/TRMM.R
@@ -8,8 +8,8 @@
 #' @param product this is the code that identifies a product, default is "3B43"
 #' @param version this is the version number, default is 7
 #' @param type this is the type of information needed, default is "precipitation.accum". Other types could be "gaugeRelativeWeighting.bin" and "relativeError.bin"
-#' @param timeExtent is a vector of dates and times for which the data should be retrieve
-#' @param bbox OPTIONAL bounding box, a list made of 4 elements: minimum longitude (lonMin), minimum latitude (latMin), maximum longitude (lonMax), maximum latitude (latMax)
+#' @param twindow is a vector of dates and times for which the data should be retrieve
+#' @param areaBox OPTIONAL bounding box, a list made of 4 elements: minimum longitude (xmin), minimum latitude (ymin), maximum longitude (xmax), maximum latitude (ymax)
 #' @param outputfileLocation file path where to save the GeoTiff
 #'
 #' @return Data is loaded as rasterbrick, then converted to a multilayer Geotiff that can
@@ -24,19 +24,19 @@
 #' @examples
 #' \dontrun{
 #'   # Define a bounding box
-#'   bbox <- list(lonMin = -3.82, latMin = 48, lonMax = -3.63, latMax = 50)
+#'   areaBox <- raster::extent(c(-3.82, -3.63, 48, 50))
 #'   twindow <- seq(as.Date("2012-01-01"), as.Date("2012-01-31"), by = "months")
 #'
 #'   TRMM(inputLocation="ftp://disc2.nascom.nasa.gov/data/TRMM/Gridded/",
 #'        product = "3B43",
 #'        version = 7,
 #'        type = "precipitation.accum",
-#'        timeExtent = twindow,
-#'        bbox = bbox,
+#'        twindow = twindow,
+#'        areaBox = areaBox,
 #'        outputfileLocation = "~/")
 #'
 #'   # or simply
-#'   TRMM(timeExtent = twindow, bbox = bbox)#'
+#'   TRMM(twindow = twindow, areaBox = areaBox)#'
 #'   plot(brick("~/trmm_acc.tif"))
 #' }
 #'
@@ -45,10 +45,9 @@ TRMM <- function(inputLocation="ftp://disc2.nascom.nasa.gov/data/TRMM/Gridded/",
                  product = "3B43",
                  version = 7,
                  type = "precipitation.accum",
-                 timeExtent = NULL,
-                 bbox = NULL,
-                 outputfileLocation = NULL
-                 ){
+                 twindow = NULL,
+                 areaBox = NULL,
+                 outputfileLocation = NULL){
 
   # Check output file location
   originalwd <- getwd()
@@ -56,45 +55,47 @@ TRMM <- function(inputLocation="ftp://disc2.nascom.nasa.gov/data/TRMM/Gridded/",
   setwd(outputfileLocation)
 
   # Check bounding box extent is within original raster extent
-  if (is.null(bbox)){
+  if (is.null(areaBox)){
     # Use TRMM max extent
-    bbox <- list(lonMin = -180, latMin = -50, lonMax = +180, latMax = +50)
+    areaBox <- raster::extent(c(-180, +180, -50, +50))
   }
-  if (bbox$lonMin < -180 | bbox$lonMin > 180) {
-    bbox$lonMin <- -180
-    message(paste("lonMin of bbox is out of the maximum extent [-180,180],",
-                  "new lonMin of bbox is modified as follows:",
-                  "lonMin =", bbox$lonMin))
+
+  if (areaBox@xmin < -180 | areaBox@xmin > 180) {
+    areaBox@xmin <- -180
+    message(paste("xmin of areaBox is out of the maximum extent",
[…]
