+++ tests/testthat/test-SEPA.R
@@ -1,17 +1,23 @@
 context("SEPA")
 
-test_that("Test catalogueSEPA function", {
-
-  skip_on_cran()
+test_that("Test full catalogueSEPA", {
 
   # Retrieve the SEPA catalogue
   x <- catalogueSEPA()
   expect_that("data.frame" %in% class(x), equals(TRUE))
   expect_that(all(dim(x) == c(830, 8)), equals(TRUE))
 
+})
+
+test_that("Test area filter with catalogueSEPA", {
+
   # Get only catchments with area above 5000 Km2
   x <- catalogueSEPA(columnName = "CatchmentAreaKm2", columnValue = ">= 5000")
   expect_that(all(dim(x) == c(12, 8)), equals(TRUE))
+
+})
+
+test_that("Test filter on the name of the river with catalogueSEPA", {
 
   # Get only catchments within river Avon
   x <- catalogueSEPA(columnName = "River", columnValue = "Avon")
@@ -20,8 +26,6 @@ test_that("Test catalogueSEPA function", {
 })
 
 test_that("Test tsSEPA function", {
-
-  skip_on_cran()
 
   # Retrieve sample data
   x <- tsSEPA("234253")

[…]
