### 2016-12-22 cvitolo/hddtools#5 @cvitolo
keep only those where "statistics" value is 1
This issue is part of this [ropensci review](https://github.com/ropensci/onboarding/issues/73).

@mdsumner suggested:
"clarify in comments that filter is bbox plus "keep only those where "statistics" value is 1"

I have added comments when I filtered using bbox but the last part of your suggestion is not clear to me, could you please clarify?
