### 2016-12-22 cvitolo/r_hddtools#5 @cvitolo
@mdsumner sorry, I now understood you refer to one of the examples in the vignette. Yes, that's now done. Thanks!
