### 2016-12-22 56281a4
Addressed suggestions from reviewer 2
