+++ tests/testthat/test-HadDAILY.R
@@ -4,6 +4,7 @@ test_that("Test HadDAILY function", {
 
   # Retrieve the full dataset (list of 11 time series)
   x <- HadDAILY()
+
   expect_that("list" %in% class(x), equals(TRUE))
   expect_that(length(x) == 11, equals(TRUE))
   expect_that(all(names(x) == c("EWP", "SEEP", "SWEP", "CEP", "NWEP", "NEEP",
