+++ tests/testthat/test-KGClimateClass.R
@@ -3,11 +3,10 @@ context("KGClimateClass")
 test_that("Test KGClimateClass function", {
 
   # Define a bounding box
-  areaBox <- raster::extent(c(-3.82, -3.63, 52.41, 52.52))
+  areaBox <- raster::extent(-3.82, -3.63, 52.41, 52.52)
   # Get climate classes
   x <- try(KGClimateClass(areaBox = areaBox), silent = TRUE)
 
-  expect_that(class(x) != "try-error", equals(TRUE))
   expect_that(all(dim(x) == c(1, 3)), equals(TRUE))
 
 })
