### 2016-12-22 #73 @cvitolo
## Responses to reviewers' questions

Many thanks to both reviewers for their extremely valuable suggestions. I realised there were some redundancies: README and vignettes contain the same examples. Therefore I decided to leave only installation and metadata in the README and left all the examples in the vignette.

I have addressed all comments and suggestions, but I have not solved open issues yet (see [here](https://github.com/cvitolo/hddtools/issues)). I hope to get this done by the end of the year.

Below are my answers to your suggestions.

### REVIEWER 1: @ledell

#### Package Metadata

> Does the package actually not work with R 3.0.1 or less, or was this just the version that was used in development and/or testing?

Thanks for pointing this out! That was just the version I used for development. However I have just tested the package on Windows with various R-Rtools versions and it seems that only stable-patched-oldrel pass all the tests. I have updated the Depends field in the DESCRIPTION file accordingly.

> Suggests: testthat
I'd add the rgdal package to at least Suggests and maybe Depends since the KGClimateClass() function depends on it.
I'd possibly consider moving the Imports to Depends for the packages that provide primary utility to the hddtools package.

Thanks, rgdal is now in Depends. I cannot move the other packages to Depends because I get a warning saying 'Depends: includes the non-default packages: ‘zoo’ ‘sp’ ‘RCurl’ ‘XML’ ‘raster’. Adding so many packages to the search path is excessive and importing selectively is preferable.' For this reason I decided to include them all in Imports.

> License: GPL-3
Are you familiar with permissive vs non-permissive licencing? A majority of commercial businesses will stray away from GPL-licensed software, so you will have a potentially wider user base if you license with a permissive license such as MIT, BSD or Apache 2.0. Just something to consider as you license your package, however if it's mostly academics who will use this package, then it probably doesn't matter too much. If you want to keep it GPL'd, you may consider with GPL (>= 2) instead of GPL-3 because it will be more compatible with other packages.

I see your point, using a permissive license would definitely widen the range of potential users but my package depends on the raster package which is under GPL-3. I thought I should apply a licence compatible with the most restrictive license of the dependencies.

> Documentation
In the description of the package, you state, "This R package is an open source project designed to facilitate non-programmatic access to a variety of online open data sources relevant for hydrologists and, more in general, environmental scientists and practitioners."
When I hear "non-programmatic access", I think of something like a GUI or website. An R API would actually be the converse -- "programmatic access". I would consider changing the wording here.

Thanks, I have removed the expression 'non-programmatic' from the documentation.

> Code style
Try to avoid using a dot in variable or function names, as in new.packages. I recommend the following edit to your README.md:
Change:
packs <- c('zoo', 'sp', 'RCurl', 'XML', 'rnrfa', 'Hmisc', 'raster',
'stringr', 'devtools')
new.packages <- packs[!(packs %in% installed.packages()[,'Package'])]
if(length(new.packages)) install.packages(new.packages)
To this:
packs <- c('zoo', 'sp', 'RCurl', 'XML', 'rnrfa', 'Hmisc', 'raster',
'stringr', 'devtools')
new_packs <- packs[!(packs %in% installed.packages()[,'Package'])]
if(length(new_packs)) install.packages(new_packs)

Thanks, this was amended.

> The convention for R code is to use spaces around all variable names and equals signs. For example:
This:
bbox <- list(lonMin=-10,latMin=48,lonMax=5,latMax=62)
Instead of this:
bbox <- list(lonMin = -10, latMin = 48, lonMax = 5, latMax = 62)

Thanks, I have applied this convention throughout. Hope I did not miss anything.

> Grammar/Spelling
README.md:
"formal request for" -> "a formal request for"
"All those operation" -> "All those operations"

Thanks, this was amended.

> LaTeX style quotes are not needed in Markdown: Update quotes ``Long-Term Mean Monthly Discharges'' (in this section)

Thanks, this was amended.

> You are missing the name of the function in this sentence, "The hddtools provides a function, called , to download "

Thanks, this was amended.

> I think it would be nice to have a few more hyperlinks to the data sources, for example, you could add a link to the GDRC or the Data60UK initiative.

Thanks, I added hyperlinks for all the data providers.

> DESCRIPTION
": Many governmental bodies" -> "Many governmental bodies"
"Especially when programmatic access to data resources is not allowed, downloading metadata catalogue, select the information needed, request datasets, de-compression, conversion, manual filtering and parsing can become rather tedious." -> "Especially when programmatic access to data resources is not allowed, downloading a metadata catalogue, selecting the information needed, requesting datasets, de-compression, conversion, manual filtering and parsing can become rather tedious."
"The package facilitate non programmatic access to various online data sources" -> "The package facilitates programmatic access to various online data sources"
"This package complements R's growing functionality in environmental web technologies to bridge the gap" -> "This package complements R's growing functionality in environmental web technologies by bridging the gap"
"data consumers and it is designed to be the starting building block" -> "data consumers. It is designed to be an initial building block"
"Facilitates discovery and handling of hydrological data, non-programmatic access to catalogues and databases." -> "Facilitates discovery and handling of hydrological data via programmatic access to catalogues and databases."

Thanks, these sentences were all amended.

> Functionality
There was a missing package when I tried to run the first example:
KGClimateClass(bbox,updatedBy='Peel')
Error:
Error in .rasterObjectFromFile(x, band = band, objecttype = "RasterLayer",  :
Cannot create RasterLayer object from this file; perhaps you need to install rgdal first
To remedy this, either add this package to the required/dependent packages in the README.md Installation section, or add a line to install the package right before using the function that depends on it. For example:
install.packages('rgdal')
KGClimateClass(bbox,updatedBy='Peel')

Thanks, rgdal is now in Depends.

> Found a possible bug:
> library(raster)
Loading required package: sp
> b <- brick('trmm_acc.tif')
> plot(b)
Error in graphics::par(old.par) :
invalid value specified for graphical parameter "pin"

I cannot reproduce your error here, have you generated the file trmm_acc.tiff using the TRMM() function before running the brick() command?
I have opened an issue [here](https://github.com/cvitolo/hddtools/issues/4).

> Is there any way you can add the size of the datasets (assuming they stay static), to the docmentation for the data downloading functions? For example, this function:
x <- catalogueData60UK()

Thanks for the suggestion, I have added the size of the dataset to the documentation.

> My plot for the following function looks quite different, I'm assuming that's because it's different data. I'd add a comment about how the plots will look different to the user, but that the following are examples of what the plots may look like.
y <- tsSEPA(hydroRefNumber = c(234253, 234174, 234305))
plot(y[[1]])

I have added a comment.

> Also, these plot all on the same plot (4 graphs on one plot). Is that intentional?
> plot(y[[1]])
> plot(y[[2]])
> plot(y[[3]])

I have changed the plot to show the three time series in different colors.

> R Documentation
Is bboxSpatialPolygon a user-level function or an internal function? It's not available to the user, apparently...
> library(hddtools)
> bbox <- list(lonMin=-180,latMin=-50,lonMax=+180,latMax=+50)
> bbSP <- bboxSpatialPolygon(bbox)
Error: could not find function "bboxSpatialPolygon"

bboxSpatialPolygon is now exported and available to the user.

> You can add linkable URLS using the \url{} command. Your links in the R docs are not currently hyperlinked.

Links in the R docs are now hyperlinked.

> In the R docs for the catalogue functions, I'd save the output to a data.frame. For example:
> # Retrieve the whole catalogue
> x <- catalogueGRDC()
> The only one that currently has this is:
> x <- catalogueSEPA()

This is now done.

> The getContent() function is not exported correctly:
>   url <- "ftp://hydrology.nws.noaa.gov/pub/gcip/mopex/US_Data/Us_438_Daily"
>   getContent(url)
Error: could not find function "getContent"

getContent is now exported and available to the user.

> The example for tsData60UK() is simply commented out instead of using the "Example Not Run" command.

Thanks, this is amended now.

> Sometimes you specify the argument name in the example and sometimes you don't. It might be best to fully specify the arguments in all the examples:
Specified:
x <- tsGRDC(stationID=1107700)
Not specified:
x <- tsMOPEX("14359000")

The arguments in all the examples are now fully specified.

> It would be good to note when a hydrometric reference number is a character string because some of the functions use numeric IDs and some use strings.

The hydrometric reference number is now always a string.

> I'd change the title/name of "SEPAcatalogue" to "Data set: The SEPA Catalogue" to make sure that it's clear that it's a dataset.

The title of SEPAcatalogue is now Data set: The SEPA Catalogue

> It might even be useful to call out any data sources as well using the term, "Data source: [name of catalogue]"

Thanks for the suggestion, the title of the various data sources is now Data source: [name of catalogue].

### REVIEWER 2: @mdsumner

> I'd like to see the examples explore the catalogue data so there's more clarity about what data are available and how to use the filters.

Thanks for the suggestion, I have now added more examples to the function documentation to clarify hw to use the filters.

> Some catalogue help pages have a full listing of their contents (e.g. SEPA) but others do not.

I have now added information to all help pages.

> Suggestions
It would be handy to see links to all the providers of data (GRDC, MOPEX, Hadley etc) in the hddtools-package doc page

Done, thanks for the suggestion.

> use raster::extent() rather than a raw list, could migrate to this default by providing an extent method for your raw lists

Thanks, hddtools now uses `raster::extent()` rather than a raw list to define a bounding box.

> note that bbox is a function in sp, so you might want a different name

Thanks, this is now called `areaBox`.

> clarify timeExtent is really an axis of steps, not a range/extent - though I think it is only used as a range (??) so range(timeExtent) would do, the seq(,by=) part is not relevant I think

I agree range(timeExtent) could suffice in the case of `tsData60UK`. However, the sequence from start to end of the time range is needed for the TRMM function because files are stored by month/day. Therefore, for a matter of consistency, I always define the timeExtent as a sequence of dates.

> include checks for sense in bounding boxes, I inadvertently tried bb <- list(lonMin=-100,latMin=43,lonMax=-50,latMax=42) which doesn't complain but will never return results (this would be provided automatically by using raster::extent)

Thanks, as mentioned above the bounding box is now defined using `raster::extent`.

> README example KGClimateClass: cvitolo/hddtools#2

I'll follow up on this in the issue (#2).

> examples use x for different calls to catalogueGRDC, please name these objects differently for interactive comparison, and do this throughout examples

Thanks for spotting this, objects in the examples are now named differently.

> clarify in comments that filter is bbox plus "keep only those where "statistics" value is 1

This is now done. Thanks!

> leaflet installation should be suggested before library(leaflet)

This was added to the vignette.

> links to GRDC and KGClimate sites, I was looking for descriptions of the columns in catalogueGRDC() and/or a link to the GRDC resources that describe this

I have added more information and a link to the help page of the GRDC catalogue.
The KGClimateClass help page contains links to the websites hosting the maps.

> README: "and, more in general, environmental scientists and practitioners." remove the "more" ?

This is now done.

> "``Long-Term Mean Monthly Discharges''" in README has funky quotations

Funky quotations are now removed.

> I suggest not using generic (and repeated) variable names like "x" and "y", use specific names for the data - so an object is standalone from examples

I have changed variable names to something more meaningful.

> use links to URLs in help pages

This is now done.

> getContent is documented but not exported

This is intentional as it is only a function used internally that wouldn't be of much use on its own.

> provide example of how to discover possible values for stationID for tsGRDC, and/or link to catalogueGRDC function in help for stationID argument
E.g.
bb <- list(lonMin=-100,latMin=42,lonMax=-50,latMax=43)
cG <- catalogueGRDC(bbox = bb)
tsg <- tsGRDC(stationID = cG$grdc_no[2])

This is now done.

> always create the bbox and timeExtent explicitly for each example, so their scope is narrow and easily runnable.

This is now done.

> show that y can be plotted directly here
hydroRefNumber <- as.numeric(as.character(x$id[1])) # e.g. 55012
y <- tsData60UK(hydroRefNumber, plotOption=FALSE)
plot(y)

This is now done.

> I ran this twice, expected the download to be cached? It downloads again
y <- tsGRDC(stationID='6122300', plotOption=TRUE)
trying URL 'ftp://ftp.bafg.de/pub/REFERATE/GRDC/ltdata/europe.zip'

No, data is not cached.

> this example doesn't link the name of the file to the read, trmm_acc.tif could be an output value of TRMM?
bbox <-
TRMM(timeExtent = timeExtent, bbox = bbox)
library(raster)
b <- brick('trmm_acc.tif')
plot(b)

This is now done.

> doc in TRMM about value is not complete "Data is loaded as rasterbrick, then converted to a multilayer Geotiff that can" also, it's really the opposite - "multilayer GeoTIFF is downloaded, loaded as a RasterBrick"

This is now amended.

> Suggest a better workflow is this, but it could go further and build the RasterBrick and return it. The raster object will record that it's linked to a file, so that seems reasonable.
trmmfile <- TRMM(timeExtent = timeExtent, bbox = bbox)
library(raster)
b <- brick(trmmfile)

This is now done.

> I get this problem, but on one machine it worked fine - I'll try to follow up
 TRMM(timeExtent = timeExtent, bbox = bbox)
latMax of bbox is out of the maximum extent [-50,50], new latMax of bbox is modified as follows: latMax = 50
Beware this function was tested on a unix machine. If you are not using a unix machine your results could be wrong. Please report any problem to cvitolodev@gmail.com, thanks!
Error in .local(.Object, ...) :
  Unable to open C:\Users\michae_sum\Documents\projects\hddtools\3B43.*.
No error
 Show Traceback
 Rerun with Debug
 Error in .rasterObjectFromFile(x, objecttype = "RasterBrick", ...) :
  Cannot create a RasterLayer object from this file.

I'll follow up on this in the issue (#3).
