### 2016-12-23 9d6b18a
All the calls to system.file are made using file.path()
