+++ DESCRIPTION
@@ -1,5 +1,4 @@
 Package: hddtools
-Type: Package
 Title: Hydrological Data Discovery Tools
 Version: 0.3.1
 Authors@R: c(person("Claudia", "Vitolo", email = "cvitolodev@gmail.com", role = c("aut", "cre")),
@@ -10,10 +9,9 @@ Maintainer: Claudia Vitolo <cvitolodev@gmail.com>
 URL: https://github.com/cvitolo/hddtools/
 BugReports: https://github.com/cvitolo/hddtools/issues
 Description: Facilitates discovery and handling of hydrological data, access to catalogues and databases.
-Depends:
-    R (>= 3.2.1), rgdal
+Depends: R (>= 3.2.1), rgdal
 Imports: zoo, sp, RCurl, XML, rnrfa, Hmisc, raster, stringr, gdata
-Suggests: testthat
+Suggests: testthat, leaflet
 License: GPL-3
 Repository: CRAN
 RoxygenNote: 5.0.1
