### 2016-12-23 73c1041
Updated KGClimateClass to print description of multiple cells
