+++ R/KGClimateClass.R
@@ -50,8 +50,8 @@ KGClimateClass <- function(areaBox = NULL, updatedBy = "Peel", verbose = FALSE){
     untar(system.file(file.path("extdata", "KOTTEK_KG.tar.gz"),
                       package = "hddtools"), exdir = td)
 
-    kgRaster <- raster::raster(paste(td, "/KOTTEK_koeppen-geiger.tiff",
-                                     sep = ""))
+    kgRaster <- raster::raster(paste0(td, "/KOTTEK_koeppen-geiger.tiff",
+                                      sep = ""))
 
     temp <- data.frame(table(raster::extract(kgRaster, bbSP)))
     temp$Class <- NA
@@ -88,7 +88,7 @@ KGClimateClass <- function(areaBox = NULL, updatedBy = "Peel", verbose = FALSE){
     untar(system.file(file.path("extdata", "PEEL_KG.tar.gz"),
                       package = "hddtools"), exdir = td)
 
-    kgRaster <- raster::raster(paste(td, "/PEEL_koppen_ascii.txt", sep = ""))
+    kgRaster <- raster::raster(paste0(td, "/PEEL_koppen_ascii.txt", sep = ""))
 
     temp <- data.frame(table(raster::extract(kgRaster, bbSP)))
     temp$Class <- NA
@@ -111,148 +111,169 @@ KGClimateClass <- function(areaBox = NULL, updatedBy = "Peel", verbose = FALSE){
   secondPart <- substr(temp$Class,2,2)
   thirdPart <- substr(temp$Class,3,3)
 
-  if (firstPart == "A"){
+  description <- vector(mode="character", length=max(length(firstPart),
+                                                           length(secondPart),
+                                                           length(thirdPart)))
 
-    description <- "A = Equatorial climates"
-    criterion <- "A = Tmin >= +18 C"
+  criterion <- vector(mode="character", length=max(length(firstPart),
+                                                         length(secondPart),
+                                                         length(thirdPart)))
 
-    if (secondPart == "f"){
-      description <- paste(description,
-                           "\nf = Equatorial rainforest, fully humid")
-      criterion <- paste(criterion,"\nf = Pmin >= 60mm")
-    }
-    if (secondPart == "m"){
-      description <- paste(description, "\nm = Equatorial monsoon")
-      criterion <- paste(criterion, "\nm = Pann >= 25*(100 - Pmin)")
-    }
-    if (secondPart == "s"){
-      description <- paste(description,
-                           "\ns = Equatorial savannah with dry summer")
-      criterion <- paste(criterion, "\ns = Pmin < 60mm in summer")
-    }
-    if (secondPart == "w"){
-      description <- paste(description,
-                           "\nw = Equatorial savannah with dry winter")
-      criterion <- paste(criterion, "\nw = Pmin < 60mm in winter")
-    }
-  }
 
-  if (firstPart == "B"){
+  for (j in 1:length(description)){
 
-    description <- "B = Arid climates"
-    criterion <- "B = Pann < 10 Pth"
+    if (firstPart[j] == "A"){
 
-    if (secondPart == "S"){
-      description <- paste(description, "\nS = Steppe climate")
-      criterion <- paste(criterion, "\nS = Pann > 5 Pth")
-    }
-    if (secondPart == "W"){
-      description <- paste(description, "\nW = Desert climate")
-      criterion <- paste(criterion, "\nW = Pann <= 5 Pth")
-    }
-  }
+      description[j] <- "A = Equatorial climates"
+      criterion[j] <- "A = Tmin >= +18 C"
 
-  if (firstPart == "C"){
+      if (secondPart[j] == "f"){
+        description[j] <- paste0(description[j],
[…]
