### 2016-12-23 cvitolo/r_hddtools#2 @cvitolo
The problem was due to some missing files in the extdata folder.
Those are now committed to the repo, and the error should not occur anymore.

This issue is now fixed.
