### 2016-12-23 cvitolo/r_hddtools#3 @cvitolo
I think the problem was due to a platform-dependent file path.
This should be now fixed but I could only test it on my ubuntu machine using the example code below.

@mdsumner Could you please check if on your machines the problem persists? Many thanks

```
# Define a bounding box
areaBox <- raster::extent(-10, 5, 48, 62)
# Define a temporal extent
twindow <- seq(as.Date("2012-01-01"), as.Date("2012-03-31"), by = "months")
# Retreive mean monthly precipitations from 3B43_V7 (based on a bounding box and time extent)
TRMMfile <- TRMM(twindow = twindow, areaBox = areaBox)
library(raster)
plot(TRMMfile)
```
