### 2016-12-23 #73 @cvitolo
I have now addressed the following issues:
- KGClimate-class problem [#2](https://github.com/cvitolo/hddtools/issues/2)
- problem with TRMM on Windows and Linux [#3](https://github.com/cvitolo/hddtools/issues/3)
- Error in graphics::par(old.par) [#4](https://github.com/cvitolo/hddtools/issues/4)

Could you please check if the problems persist? Many thanks!
