### 2016-12-25 86e11c1
Modified the website-reference to separate the functions
