### 2016-12-27 7bacfc3
Added JOSS paper under review to website (articles section)
