+++ vignettes/hddtools_vignette.Rmd
@@ -1,7 +1,7 @@
 ---
 title: "hddtools: Hydrological Data Discovery Tools"
 author: "Claudia Vitolo"
-date: "22 December 2016"
+date: "`r Sys.Date()`"
 output: html_document
 ---
 
