+++ _pkgdown.yml
@@ -6,8 +6,8 @@ navbar:
     href: reference/index.html
   - text: Articles
     menu:
-    - text: JOSS paper
-      href: http://dx.doi.org/10.21105/joss.00052
+    - text: JOSS paper (under review)
+      href: https://github.com/openjournals/joss-reviews/issues/56
     - text: Abstract/Poster presented at AGU 2014
       href: http://adsabs.harvard.edu/abs/2014AGUFM.H13H1209V
     - text: Vignette
