### 2016-12-27 b440c73
Updated JOSS paper with latest citations
