+++ inst/paper/paper.bib
@@ -1,10 +1,10 @@
-@misc{hddtools-archive,
-  author       = {Claudia Vitolo},
-  title        = {hddtools: Hydrological Data Discovery Tools},
-  month        = Sep,
-  year         = 2016,
-  doi          = {10.5281/zenodo.61570},
-  url          = {http://dx.doi.org/10.5281/zenodo.61570}
+@Manual{hddtoolsCRAN,
+    author = {Claudia Vitolo},
+    title = {hddtools: Hydrological Data Discovery Tools},
+    year = {2016},
+    note = {R package version 0.3.0},
+    url = {https://CRAN.R-project.org/package=hddtools},
+    doi = {10.5281/zenodo.61570},
 }
 
 @Manual{R-base,
@@ -30,11 +30,61 @@
 }
 
 @Manual{rnrfa,
-  author = {Claudia Vitolo and
-            Matthew Fry and
-            Wouter Buytaert},
+  author = {Claudia Vitolo and Matthew Fry and Wouter Buytaert},
   title = {rnrfa: UK National River Flow Archive Data from R},
   year = 2015,
   note = {R package version 0.5.3},
   url = {https://CRAN.R-project.org/package=rnrfa}
+}
+
+@Manual{fuseGitHub,
+    author = {Claudia Vitolo and Peter Wells and Martin Dobias and Wouter Buytaert},
+    title = {fuse: Framework for Understanding Structural Errors},
+    year = {2012},
+    note = {R package version 3.0.0},
+    url = {https://github.com/cvitolo/r_fuse},
+    doi = {http://dx.doi.org/10.5281/zenodo.14005},
+  }
+
+@Article{fuseJOSS,
+  author = {Claudia Vitolo and Peter Wells and Martin Dobias and Wouter Buytaert},
+  title = {fuse: An R package for ensemble Hydrological Modelling},
+  journal = {The Journal of Open Source Software},
+  doi = {10.21105/joss.00052},
+  url = {http://dx.doi.org/10.21105/joss.00052},
+  year  = {2016},
+  month = {dec},
+  publisher = {The Open Journal},
+  volume = {1},
+  number = {8}
+}
+
+@Manual{topmodel,
+    title = {topmodel: Implementation of the hydrological model TOPMODEL in R},
+    author = {Wouter Buytaert},
+    year = {2011},
+    note = {R package version 0.7.2-2},
+    url = {https://CRAN.R-project.org/package=topmodel},
+  }
+
+@Manual{hydromad,
+    title = {hydromad: Hydrological Model Assessment and Development},
+    author = {Felix Andrews and Joseph Guillaume},
+    year = {2016},
+    note = {R package version 0.9-25},
+    url = {http://hydromad.catchment.org/},
+  }
+
+@article{Andrews20111171,
+title = "An open software environment for hydrological model assessment and development ",
+journal = "Environmental Modelling & Software ",
+volume = "26",
+number = "10",
+pages = "1171 - 1185",
+year = "2011",
+note = "",
+issn = "1364-8152",
+doi = "http://dx.doi.org/10.1016/j.envsoft.2011.04.006",
+url = "http://www.sciencedirect.com/science/article/pii/S1364815211001046",
+author = "F.T. Andrews, and B.F.W. Croke, and A.J. Jakeman",
 }
