### 2016-12-27 cvitolo/r_hddtools#4 @mdsumner
Seems fine now, here on Windows.
