### 2016-12-28 cvitolo/r_hddtools#4 @cvitolo
@mdsumner  many thanks!

@ledell is this working for you too?
