### 2016-12-28 45152f9
Fixed typo in hddtools-package.R
