### 2016-12-30 cvitolo/r_hddtools#4 @ledell
@cvitolo Yes, this is working on my end (on a mac).
