### 2017-01-02 cvitolo/r_hddtools#4 @cvitolo
Great, thanks! I'm closing this now :)
