### 2017-01-03 #73 @cvitolo
@sckott @ledell @mdsumner First of all happy new year!

I have done my best to address all the comments and issues. Please let me know how you want to proceed. Many thanks for your insights and comments, they greatly improved this package!
