### 2017-01-03 #73 @noamross
Thanks for the updates, @cvitolo! @ledell and @mdsumner, can you look at the updates to see if all your concerns have been addressed?
