### 2017-01-07 655cad2
Added note of ROPENSCI and JOSS reviews
