### 2017-01-09 cvitolo/r_hddtools#3 @mdsumner
Fine on Windows now, looks good!
