### 2017-01-12 #73 @cvitolo
Many thanks @ledell and @mdsumner, your suggestions greatly improved the package!

@noamross Is there anything else needed on my side? Maybe create a new release? Many thanks!
