### 2017-01-16 #73 @noamross
Approved!  Remaining steps:

- You need `rmarkdown` and `knitr` in `Suggests`
- You need an entry in your `DESCRIPTION` file of: `VignetteBuilder: knitr`
- I've added you to an hddtools team for the rOpenSci repository. Accept and transfer the repo (under "Settings") to rOpenSci, and then update any links for CI, coverage, and documentation.
- Create a new release and provide the updated Zenodo DOI if needed.  I'll see if we can re-start the JOSS process without the need to submit again.
- Let me know if you are interested in doing a blog post about your package on the rOpenSci blog, either long-form (https://ropensci.org/blog/) or short (https://ropensci.org/tech-notes/)
