### 2017-01-16 39aae0d
Package accepted by rOpenSci, updated to v0.4 ready to transfer.
