+++ _pkgdown.yml
@@ -16,7 +16,7 @@ navbar:
     href: news/index.html
   right:
   - icon: fa-github fa-lg
-    href: https://github.com/cvitolo/hddtools/
+    href: https://github.com/ropensci/hddtools/
 
 reference:
 - title: Get catalogues
