+++ DESCRIPTION
@@ -6,8 +6,8 @@ Authors@R: c(person("Claudia", "Vitolo", email = "cvitolodev@gmail.com", role = 
              person("Erin", "Le Dell", role = c("ctb"), comment = "Erin reviewed the package for rOpenSci, see https://github.com/ropensci/onboarding/issues/73"),
              person("Michael", "Sumner", role = c("ctb"), comment = "Michael reviewed the package for rOpenSci, see https://github.com/ropensci/onboarding/issues/73"))
 Maintainer: Claudia Vitolo <cvitolodev@gmail.com>
-URL: https://github.com/cvitolo/hddtools/
-BugReports: https://github.com/cvitolo/hddtools/issues
+URL: https://github.com/ropensci/hddtools
+BugReports: https://github.com/ropensci/hddtools/issues
 Description: Facilitates discovery and handling of hydrological data, access to catalogues and databases.
 Depends: R (>= 3.2.1), rgdal
 Imports: zoo, sp, RCurl, XML, rnrfa, Hmisc, raster, stringr, gdata
