### 2017-01-16 #73 @cvitolo
Hi @noamross,

I have updated the DESCRIPTION file as you requested, transferred the repository to rOpenSci, created a new release and updated the links.

Unfortunately, the new repo ropensci/hddtools is not appearing in Zenodo and I cannot generate a new DOI. I suppose you need to allow Zenodo application to access the new repository. Would it be possible to do that please?

Also, I have a website in the docs folder, would you mind to give me access to the settings of this new repo so that I can set up the website to be rendered in github?

Happy to write a blog post about my package.

Many thanks!

Claudia
