### 2017-01-16 #73 @noamross
You should have admin access to the ropensci repository. Can you tell me whether you are able to give Zenodo access?
