+++ .travis.yml
@@ -22,6 +22,7 @@ before_install:
   - ./configure --prefix=/usr
   - make
   - sudo make install
+  - cd /home/travis/build/ropensci/hddtools
 
 r_github_packages:
   - jimhester/covr
