### 2017-01-16 c2876eb
Updated to v0.5
