### 2017-01-16 be938c1
Updated DOI after release v0.5
