+++ .travis.yml
@@ -2,7 +2,6 @@
 
 language: r
 sudo: required
-cache: packages
 
 before_install:
   - curl -OL https://raw.githubusercontent.com/metacran/r-builder/master/pkg-build.sh
