### 2017-01-16 d26b013
Pre-update pkgs to avoid Travis dependency error
The Travis failure appears to be a known issue that should be transient due to Ubuntu package update schedules.  It is described with this workaround here: https://github.com/travis-ci/travis-ci/issues/6850
