### 2017-01-16 10bf38d
Added to travis:   Rscript -e 'update.packages(ask = FALSE)'
