+++ R/GRDC.R
@@ -38,7 +38,8 @@ catalogueGRDC <- function(areaBox = NULL,
                           columnName = NULL, columnValue = NULL,
                           mdDescription = FALSE, useCachedData = TRUE){
 
-  theurl <- "http://www.bafg.de/GRDC/EN/02_srvcs/21_tmsrs/211_ctlgs/GRDC_Stations.zip?__blob=publicationFile"
+  theurl <- paste0("http://www.bafg.de/GRDC/EN/02_srvcs/21_tmsrs/211_ctlgs/",
+                   "GRDC_Stations.zip?__blob=publicationFile")
 
   if (useCachedData == TRUE | RCurl::url.exists(theurl) == FALSE){
 
@@ -175,10 +176,10 @@ catalogueGRDC <- function(areaBox = NULL,
 #' }
 #' \item{\strong{mddPerMonth}}{This is a table containing mean daily discharges for each month over the entire period (12 records covering max. n years. Calculated only for months with less then or equal to 10 missing days). It is made of 7 columns which description is as follows:}\itemize{
 #'   \item LQ:    lowest monthly discharge of the given month in the entire period
-#'   \item year:  associated year of occurrence (only the first occurence is listed)
+#'   \item year:  associated year of occurrence (only the first occurrence is listed)
 #'   \item MQ:    mean discharge from all monthly discharges of the given month in the entire period
 #'   \item HQ:    highest monthly discharge of the given month in the entire period
-#'   \item year:  associated year of occurrence (only the first occurence is listed)
+#'   \item year:  associated year of occurrence (only the first occurrence is listed)
 #'   \item std:   standard deviation of all monthly discharges of the given month in the entire period
 #'   \item n:     number of available daily values used for computation
 #' }
